+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Change Log (vs-script-commands)
 
+## 1.2.0 (January 23rd, 2017; suppressArguments)
+
+* added `arguments` and `suppressArguments` to [ScriptCommandExecutorArguments](https://mkloubert.github.io/vs-script-commands/interfaces/_contracts_.scriptcommand.html)
+
+
 ## 1.1.0 (January 23rd, 2017; extended API)
 
-- extended [ScriptCommandExecutorArguments](https://mkloubert.github.io/vs-script-commands/interfaces/_contracts_.scriptcommandexecutorarguments.html) interface
+* extended [ScriptCommandExecutorArguments](https://mkloubert.github.io/vs-script-commands/interfaces/_contracts_.scriptcommandexecutorarguments.html) interface
+
 
 ## 1.0.0 (January 23rd, 2017; first release)
 
-- first official release
+* first official release